Add files via upload
The Olympic Games are the biggest Sporting Event in the world. Inspired by the ancient Olympic Games of the Greek, they provide an extremely competitive platform for nations to showcase their sporting talents. Using the data provided, I will try to extract some interesting facts, such as the youngest and oldest participants, the participants who appeared the most times, the ratio of male and female participants, the most popular sports, the countries with the most medals won etc. I will visually present this data, using the Seaborn library. Seaborn is a Python data visualization library based on matplotlib. It provides a high-level interface for drawing attractive and informative statistical graphics.